refactor: 정적 펙토리 적용

+++ src/main/java/org/example/food/service/VideoServiceImpl.java
@@ -11,11 +11,11 @@ import org.example.food.exception.VideoExceptionType;
 import org.example.food.repository.LikeRepository;
 import org.example.food.repository.VideoQueryRepository;
 import org.example.food.repository.VideoRepository;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.domain.SliceImpl;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.data.domain.*;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -26,73 +26,57 @@ import java.util.stream.Collectors;
 public class VideoServiceImpl implements VideoService {
 
     private final VideoRepository videoRepository;
-    private final RestaurantService restaurantService; // Restaurant 관련 로직 분리
+    private final RestaurantService restaurantService;
     private final VideoQueryRepository videoQueryRepository;
     private final LikeRepository likeRepository;
 
-    // Video -> VideoResDto 변환
+    @Override
+    public Slice<VideoResDto> getAllVideos(Pageable pageable, User user) {
+        return getVideos(videoQueryRepository.findAllVideosWithPagination(pageable), user, pageable);
+    }
+
+    @Override
+    public Slice<VideoResDto> getNearbyVideos(double userLat, double userLon, double radius, Pageable pageable, User user) {
+        return getVideos(videoQueryRepository.findVideosByLocationWithPagination(userLat, userLon, radius, pageable), user, pageable);
+    }
+
+    private Slice<VideoResDto> getVideos(List<Video> videos, User user, Pageable pageable) {
+        List<VideoResDto> content = videos.stream()
+                .map(video -> toVideoDto(video, user))
+                .collect(Collectors.toList());
+
+        boolean hasNext = content.size() > pageable.getPageSize();
+        if (hasNext) {
+            content.remove(pageable.getPageSize());
+        }
+        return new SliceImpl<>(content, pageable, hasNext);
+    }
+
     private VideoResDto toVideoDto(Video video, User user) {
         if (video == null) {
             return null;
         }
-        boolean isLike = likeRepository.findByUserAndVideo(user, video).isPresent();  // 좋아요 여부 확인
 
-        return VideoResDto.builder()
-                .videoId(video.getId())
-                .url(video.getUrl())
-                .content(video.getContent())
-                .category(video.getCategory())
-                .restaurant(video.getRestaurant().getName())
-                .restaurantId(video.getRestaurant().getId())
-                .userNickname(video.getUser().getNickname())
-                .likeCount(video.getLikeCount())
-                .isLike(isLike)
-                .build();
-    }
-
-    // VideoReqDto -> Video 변환
-    private Video toEntity(VideoReqDto dto) {
-        if (dto == null) {
-            return null;
-        }
-        Restaurant restaurant = restaurantService.findOrCreateRestaurant(dto);
-        return Video.builder()
-                .url(dto.getUrl())
-                .content(dto.getContent())
-                .category(dto.getCategory())
-                .restaurant(restaurant)
-                .build();
-    }
-
-    @Override
-    public Slice<VideoResDto> getAllVideos(Pageable pageable, User user) {
-        List<Video> videos = videoQueryRepository.findAllVideosWithPagination(pageable);
-
-        List<VideoResDto> content = videos.stream()
-                .map(video -> toVideoDto(video, user))  // 현재 사용자를 함께 전달
-                .collect(Collectors.toList());
-
-        boolean hasNext = content.size() > pageable.getPageSize();
-
-        if (hasNext) {
-            content.remove(pageable.getPageSize());  // 페이지 사이즈 초과시 마지막 요소 제거
-        }
-
-        return new SliceImpl<>(content, pageable, hasNext);
+        boolean isLike = user != null && likeRepository.findByUserAndVideo(user, video).isPresent(); // 좋아요 여부 확인
+        return VideoResDto.fromVideo(video, isLike);
     }
 
     @Override
     @Transactional
     public Long createVideo(VideoReqDto videoReqDto, User user) {
-        Video video = toEntity(videoReqDto);
-        video.setUser(user);
+        Restaurant restaurant = restaurantService.findOrCreateRestaurant(videoReqDto);
+        Video video = Video.toEntity(videoReqDto, user, restaurant);
         videoRepository.save(video);
         return video.getId();
     }
 
+
     @Override
     @Transactional
     public void deleteVideo(Long id) {
+        if (!videoRepository.existsById(id)) {
+            throw new VideoException(VideoExceptionType.NOT_FOUND_VIDEO);
+        }
         videoRepository.deleteById(id);
     }
 
@@ -100,22 +84,5 @@ public class VideoServiceImpl implements VideoService {
     public Video findVideoById(Long id) {
         return videoRepository.findById(id)
                 .orElseThrow(() -> new VideoException(VideoExceptionType.NOT_FOUND_VIDEO));
-    }
-
-    @Override
-    public Slice<VideoResDto> getNearbyVideos(double userLat, double userLon, double radius, Pageable pageable, User user) {
-        List<Video> videos = videoQueryRepository.findVideosByLocationWithPagination(userLat, userLon, radius, pageable);
-
-        List<VideoResDto> content = videos.stream()
-                .map(video -> toVideoDto(video, user))
-                .collect(Collectors.toList());
-
-        boolean hasNext = content.size() > pageable.getPageSize();
-
-        if (hasNext) {
-            content.remove(pageable.getPageSize());  // 페이지 사이즈 초과시 마지막 요소 제거
-        }
-
-        return new SliceImpl<>(content, pageable, hasNext);
     }
 }
